# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recruitment Dashboard API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/candidates": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_api_candidates_get",
        "parameters": [
          {
            "name": "skills",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skills"
            }
          },
          {
            "name": "min_exp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Exp"
            }
          },
          {
            "name": "max_exp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Exp"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "skill_experience",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skill Experience"
            }
          },
          {
            "name": "role_experience",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Experience"
            }
          },
          {
            "name": "skill_match_mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "any"
              ],
              "type": "string",
              "default": "all",
              "title": "Skill Match Mode"
            }
          },
          {
            "name": "comment_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "desc",
                "asc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Comment Order"
            }
          },
          {
            "name": "requirement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRead"
                  },
                  "title": "Response List Candidates Api Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}": {
      "patch": {
        "tags": [
          "Candidates"
        ],
        "summary": "Update Candidate",
        "operationId": "update_candidate_api_candidates__candidate_id__patch",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/interviews": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Create Candidate Interview",
        "operationId": "create_candidate_interview_api_candidates__candidate_id__interviews_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/interviews/{interview_id}": {
      "patch": {
        "tags": [
          "Candidates"
        ],
        "summary": "Update Candidate Interview",
        "operationId": "update_candidate_interview_api_candidates__candidate_id__interviews__interview_id__patch",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/matching/{requirement_id}/candidates/{candidate_id}": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Run Candidate Matching",
        "operationId": "run_candidate_matching_api_matching__requirement_id__candidates__candidate_id__post",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResultRead"
                  },
                  "title": "Response Run Candidate Matching Api Matching  Requirement Id  Candidates  Candidate Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateEducationRead": {
        "properties": {
          "institute": {
            "type": "string",
            "title": "Institute"
          },
          "degree_name": {
            "type": "string",
            "title": "Degree Name"
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "year_of_passing": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Of Passing"
          },
          "gpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpa"
          }
        },
        "type": "object",
        "required": [
          "institute",
          "degree_name"
        ],
        "title": "CandidateEducationRead"
      },
      "CandidateExperienceRead": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "skills_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Used"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "CandidateExperienceRead"
      },
      "CandidateMatchView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "current_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Company"
          },
          "resume_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Url"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CandidateMatchView"
      },
      "CandidateProjectRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "skills_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Used"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CandidateProjectRead"
      },
      "CandidateRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "current_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Company"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "highest_degree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Degree"
          },
          "year_of_passing": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Of Passing"
          },
          "gpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpa"
          },
          "resume_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Url"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "requirement_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "not_applied",
                  "new",
                  "processing",
                  "rejected",
                  "hired"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Status"
          },
          "interview_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Date"
          },
          "interview_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Time"
          },
          "interviews": {
            "items": {
              "$ref": "#/components/schemas/InterviewRead"
            },
            "type": "array",
            "title": "Interviews"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "structured_profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Profile"
          },
          "skill_profiles": {
            "items": {
              "$ref": "#/components/schemas/CandidateSkillRead"
            },
            "type": "array",
            "title": "Skill Profiles"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/CandidateExperienceRead"
            },
            "type": "array",
            "title": "Experiences"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/CandidateProjectRead"
            },
            "type": "array",
            "title": "Projects"
          },
          "educations": {
            "items": {
              "$ref": "#/components/schemas/CandidateEducationRead"
            },
            "type": "array",
            "title": "Educations"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CandidateRead"
      },
...
      "CandidateSkillRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "context": {
            "type": "string",
            "enum": [
              "primary",
              "secondary",
              "project",
              "mentioned"
            ],
            "title": "Context"
          },
          "experience_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Months"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          }
        },
        "type": "object",
        "required": [
          "name",
          "context"
        ],
        "title": "CandidateSkillRead"
      },
      "CandidateUpdate": {
        "properties": {
          "interview_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Date"
          },
          "interview_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Time"
          }
        },
        "type": "object",
        "title": "CandidateUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewCreate": {
        "properties": {
          "interview_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Date"
          },
          "interview_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Time"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "InterviewCreate"
      },
      "InterviewRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Round"
          },
          "interview_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Date"
          },
          "interview_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Time"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "InterviewRead"
      },
      "InterviewUpdate": {
        "properties": {
          "interview_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Date"
          },
          "interview_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Time"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "InterviewUpdate"
      },
...
      "MatchResultRead": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateMatchView"
          },
          "requirement": {
            "$ref": "#/components/schemas/RequirementMatchView"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_applied",
              "new",
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "candidate",
          "requirement",
          "score",
          "reason",
          "status"
        ],
        "title": "MatchResultRead"
      },
...
      "RequirementMatchView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "RequirementMatchView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}